Add a deps.edn to ease local development
Projects that use tools-deps and want to depend on a specific commit sha
in the crucible project need a deps.edn present.
Ideally project.clj and deps.end dependencies should sync automatically.
The is some activity in that area, for example
https://github.com/RickMoynihan/lein-tools-deps could fix this. For now
we will have to keep it in sync. Switching to tools-deps is also an
option...