Condition policy (#172)
* Adds key override for cool-down key

The resulting json key from cool-down spec is Cooldown and not CoolDown. I added an override to stay compatible with the current keys. However we could just correct the spec key to be ::cooldown. https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-as-group.html#cfn-as-group-cooldown

* Don’t try to guess the policy

There is an issue with trying to us s/or to determine the policy in that the structure of the data might be common between more than one policy. Example of this is the condition policy and depends-on. Both accept a keyword. So s/or will set the it as  condition or depends-on based on which one of those policies appear first in the s/or ::policy. The solution is to use the ::policies spec that makes use of the s/keys function. Then have the policy function used in the resource return an explicit policy spec.